Dashboard: Thêm Linechart
